+++ lib/ui/screen-components/protected/cuisines/view/header/index.tsx
@@ -2,35 +2,12 @@ import HeaderText from '@/lib/ui/useable-components/header-text';
 import TextIconClickable from '@/lib/ui/useable-components/text-icon-clickable';
 import { ICuisineScreenHeaderProps } from '@/lib/utils/interfaces/cuisine.interface';
 import { faAdd } from '@fortawesome/free-solid-svg-icons';
-import { Sidebar } from 'primereact/sidebar';
-import CuisineForm from './CuisineForm';
 
 export default function CuisineScreenHeader({
-  setVisible,
-  setCuisines,
-  cuisines,
-  handleAddCuisineLocally,
-  visible,
-  setIsEditing,
-  isEditing,
   handleButtonClick,
 }: ICuisineScreenHeaderProps) {
   return (
     <div>
-      <Sidebar
-        visible={visible}
-        onHide={() => setVisible(false)}
-        position="right"
-      >
-        <CuisineForm
-          setVisible={setVisible}
-          setCuisines={setCuisines}
-          addCuisineLocally={handleAddCuisineLocally}
-          isEditing={isEditing}
-          setIsEditing={setIsEditing}
-          cuisines={cuisines}
-        />
-      </Sidebar>
       <div className="flex justify-between items-center p-2 w-full">
         <HeaderText text="Cuisines" />
         <TextIconClickable
