--- lib/ui/screen-components/protected/layout/super-admin-layout/management/notifications/NotificationTable.tsx
@@ -1,3 +1,0 @@
-export default function NotificationTable() {
-  return <div></div>;
-}
