--- lib/ui/screen-components/protected/layout/super-admin-layout/management/notifications/SendNotification.tsx
@@ -1,133 +1,0 @@
-//queries
-import { SEND_NOTIFICATION_USER } from '@/lib/api/graphql';
-
-//contexts
-import { ToastContext } from '@/lib/context/toast.context';
-
-//icons
-
-//prime react
-import { Button } from 'primereact/button';
-
-//components
-import CustomTextAreaField from '@/lib/ui/useable-components/custom-text-area-field';
-import CustomTextField from '@/lib/ui/useable-components/input-field';
-
-//hooks & react interfaces
-import { useMutation } from '@apollo/client';
-import { ProgressSpinner } from 'primereact/progressspinner';
-import {
-  ChangeEvent,
-  Dispatch,
-  FormEvent,
-  SetStateAction,
-  useContext,
-  useState,
-} from 'react';
-
-export default function SendNotification({
-  setVisible,
-}: {
-  setVisible: Dispatch<SetStateAction<boolean>>;
-}) {
-  //toast
-  const { showToast } = useContext(ToastContext);
-
-  //states
-  const [notificationData, setNotificationData] = useState({
-    title: '',
-    body: '',
-  });
-
-  //mutation
-  const [sendNotificationUser, { loading }] = useMutation(
-    SEND_NOTIFICATION_USER
-  );
-
-  //form change
-  const handleFormChange = (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const { value, name } = e.target;
-    setNotificationData((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-  };
-
-  //handle
-
-  //form submit
-  const handleFormSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    try {
-      await sendNotificationUser({
-        variables: {
-          notificationTitle: notificationData.title,
-          notificationBody: notificationData.body,
-        },
-      });
-      setNotificationData({
-        body: '',
-        title: '',
-      });
-      return showToast({
-        title: 'New Notification',
-        type: 'success',
-        message: 'Notification has been sent successfully',
-        duration: 2500,
-      });
-    } catch (err) {
-      setVisible(true);
-      showToast({
-        title: 'New Notification',
-        type: 'error',
-        message: 'Something went wrong',
-        duration: 2500,
-      });
-      return console.log(err);
-    }
-  };
-
-  return (
-    <div>
-      <form className="flex flex-col gap-8" onSubmit={handleFormSubmit}>
-        <h2 className="font-bold mb-3 text-xl">Send a Notification</h2>
-        <CustomTextField
-          value={notificationData.title}
-          onChange={handleFormChange}
-          name="title"
-          className="w-full py-2 px-1 text-sm"
-          showLabel={true}
-          placeholder="Title"
-          type="text"
-        />
-        <CustomTextAreaField
-          value={notificationData.body}
-          onChange={handleFormChange}
-          showLabel={true}
-          label="Description"
-          name="body"
-          placeholder="Add description here"
-          className="w-full text-sm"
-          rows={5}
-        />
-        <Button
-          type="submit"
-          className="bg-black text-white p-2 w-32 right-0 self-end flex items-center justify-center hover:bg-[#000000d8]"
-        >
-          {loading ? (
-            <ProgressSpinner
-              className="w-6 h-6 items-center self-center m-0 p-0"
-              strokeWidth="5"
-              style={{ fill: 'white', accentColor: 'white' }}
-              color="white"
-            />
-          ) : (
-            'Send'
-          )}
-        </Button>
-      </form>
-    </div>
-  );
-}
